Link Doom into the firmware and boot it as far as R_Init
The engine now links and runs on the badge. It maps 3,362,373 bytes of WAD
straight from flash, reads its 1137-lump directory, identifies itself as
DOOM Shareware, and gets through W_Init, I_Init and M_Init before the zone
heap runs out during texture setup.

Getting there:

  - main.c actually calls D_DoomMain. Until now nothing referenced it, so
    the linker was quietly discarding the entire engine and the build
    looked green while containing no Doom at all.
  - D_FindIWAD is bypassed. It probes for files by name, and there is no
    filesystem; the badge wad_file_class ignores the path anyway. gamemode
    is still derived from the lumps, so a different WAD still works.
  - drone, net_client_connected and snd_musicdevice are defined false in
    the stubs. Their definitions lived in the net client and sound layer,
    and with them false every branch takes the single-player path.

Freeing the DRAM to get this far:

  MAXVISPLANES 128 -> 24, MAXOPENINGS 320*64 -> 320*8,
  MAXDRAWSEGS 256 -> 96, MAXVISSPRITES 128 -> 48
  viewangletox int -> short: 4096 entries holding values that span
    -1..321, so 16,384 bytes became 8,192
  FreeRTOS and the SPI driver moved out of IRAM, which on this chip is the
    same 400 KB as DRAM: IRAM .text 49,984 -> 37,216
  DMA chunks 16 rows -> 8

  engine RAM 277,160 -> 170,262;  image DRAM remaining 98,788 -> 130,180

I_ZoneBase was also wrong in a way worth recording. It sized the zone by
subtracting a reserve from the largest free block -- but the reserve exists
for W_AddFile's lump directory, which needs its own contiguous block, and
taking it out of the same block it is sizing against helps nothing. That is
what "Couldn't realloc lumpinfo" meant: not a zone failure at all. It now
grabs the largest block and hands it back if too little contiguous space
survives elsewhere.

The remaining blocker is purely zone size. 113,252 bytes are free after the
framebuffer but only 49,152 are contiguous, and R_Init needs more than that.

+++ firmware/components/doom/src/r_main.c
@@ -90,7 +90,10 @@ angle_t			clipangle;
 // maps the visible view angles to screen X coordinates,
 // flattening the arc to a flat projection plane.
 // There will be many angles mapped to the same X. 
-int			viewangletox[FINEANGLES/2];
+// 4096 entries, each holding a screen column clamped to -1..viewwidth+1.
+// As int that is 16,384 bytes; as short it is 8,192, and the range fits with
+// three orders of magnitude to spare.
+short			viewangletox[FINEANGLES/2];
 
 // The xtoviewangleangle[] table maps a screen pixel
 // to the lowest viewangle that maps back to x ranges
